pyproject.toml
[tool.poetry]
name = "test-task-wht-agency"
version = "0.1.0"
description = ""
authors = ["Artem Kozatskyi <artem.kozatskyi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
dependency-injector = "^4.41.0"
django-annoying = "^0.10.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
